+++ src/tree/v2/application/registerResourceGroupsTreeV2.ts
@@ -2,7 +2,7 @@ import * as vscode from 'vscode';
 import { ApplicationResourceProviderManager } from '../../../api/v2/ApplicationResourceProviderManager';
 import { localize } from './../../../utils/localize';
 import { ApplicationResourceGroupingManager } from './ApplicationResourceGroupingManager';
-import { createBranchDataItemFactory } from '../BranchDataItem';
+import { createBranchDataItemFactory } from '../BranchDataProviderItem';
 import { createGroupingItemFactory } from './GroupingItem';
 import { ApplicationResourceBranchDataProviderManager } from './ApplicationResourceBranchDataProviderManager';
 import { ResourceGroupsItemCache } from '../ResourceGroupsItemCache';
